updated documentation, fixed 64bit size_t compile error in ex program
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@20408 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 4a6505a19884c7ac03098e45c980c8f33f87dba9

+++ sandbox/descriptors_3d/examples/interest_point_example.cpp
@@ -60,7 +60,7 @@ int main()
   createPointCloud(data);
   cloud_kdtree::KdTreeANN data_kdtree(data);
   cv::Vector<const robot_msgs::Point32*> interest_pts(data.pts.size());
-  for (unsigned int i = 0 ; i < data.pts.size() ; i++)
+  for (size_t i = 0 ; i < data.pts.size() ; i++)
   {
     interest_pts[i] = &(data.pts[i]);
   }
